refactor(gui): rename utilitybestvfsavailablemode to job
Renamed utilitybestvfsavailablemode.{h,cpp} to utilitybestvfsavailablemodejob.{h,cpp} for naming consistency.
Updated all references in CMakeLists.txt and guijobfactory.cpp.
No functional changes: logic for determining best VFS mode remains unchanged.
